Object / Response Parsing Updates
Ran into an issue where Spoke chokes when told to consume the results of
it's own BackFillClockMessages API method. BackFillClockMessages returns
a list of clock messages, but when none are missing it just returns an
empty set of object brackets like so: []

The http response objects use a deserializer to parse this into a
dynamic object. The dynamic object then becomes an empty array rather
than remaining a dynamic object (since the type becomes identifiable at
the time of execution). Attempting to then access the 'StatusCode'
property of the now-an-array object crashes and burns. This causes a
situation where we get bad subscription spam forever, increasing load
and using up threads.

This fix does two things:
First, I normalized all of the Spoke api methods (public and private, as
private may be exposed as public on individual implementations) to use a
response object. This means even an empty object will still turn out to
be a valid object with an empty array of values on that key. This
provides a more consistent interface for consumers interfacing with
Spoke.

Second, I implemented an Expando dynamic object for the http response
handlers. The values are now stored in the obj.ResponseValue instead of
the values from the response becoming the object explicitly. This means
even junk, trash, or poorly formatted responses won't crash the HTTP
handlers.

This update might be incompatible with previous Spoke integrations that
relied on non-object responses from the fixed API calls. I've tested
this internally to Spoke and  made the necessary adjustments. I'm not
aware of anybody programmatically accessing Spoke in a way that this
update would affect, but it's worth thinking about.